+++ app/com/atlassian/connect/play/java/service/AcHostService.java
@@ -1,6 +1,9 @@
 package com.atlassian.connect.play.java.service;
 
 import com.atlassian.connect.play.java.AcHost;
+import com.atlassian.fugue.Option;
+
+import java.util.List;
 
 import static play.libs.F.Promise;
 
@@ -18,12 +21,23 @@ public interface AcHostService {
      */
     Promise<String> fetchPublicKeyFromRemoteHost(AcHost acHost);
 
+    Promise<Void> registerHost(AcHost acHost);
+
     /**
      * Registers a new host, or updates the details from an existing host
      *
      *
-     * @param acHost
+     * @param clientKey
+     * @param  baseUrl
+     * @param publicKey
+     * @param sharedSecret
+     * @param name
      * @return
      */
-    Promise<Void> registerHost(AcHost acHost);
+    Promise<Void> registerHost(String clientKey, String baseUrl, String publicKey, String sharedSecret, String name);
+
+    Option<AcHost> findByKey(String consumerKey) throws Throwable;
+
+    List<AcHost> all() throws Throwable;
+
 }