+++ app/com/atlassian/connect/play/java/service/AcHostServiceImpl.java
@@ -4,11 +4,15 @@ import com.atlassian.connect.play.java.AcHost;
 import com.atlassian.connect.play.java.auth.InvalidAuthenticationRequestException;
 import com.atlassian.connect.play.java.auth.MismatchPublicKeyException;
 import com.atlassian.connect.play.java.auth.PublicKeyVerificationFailureException;
-import com.google.common.annotations.VisibleForTesting;
+import com.atlassian.fugue.Option;
 import com.google.common.base.Objects;
+import com.google.common.base.Supplier;
 import org.w3c.dom.Document;
 import org.w3c.dom.NodeList;
 
+import java.util.List;
+
+import static com.atlassian.fugue.Option.none;
 import static org.apache.commons.lang.StringUtils.stripToNull;
 import static play.libs.F.Function;
 import static play.libs.F.Promise;
@@ -52,6 +56,36 @@ public class AcHostServiceImpl implements AcHostService {
 
     @Override
     public Promise<Void> registerHost(final AcHost acHost) {
+        return registerHost(acHost.getKey(), acHost.getBaseUrl(), acHost.getPublicKey(), acHost.getSharedSecret(),
+                acHost.getName());
+    }
+
+    @Override
+    public Promise<Void> registerHost(final String clientKey, final String baseUrl, String publicKey, String sharedSecret, String name) {
+        // TODO: The consequence of this is that we will overwrite registrations each time. Is that what we want?
+        Option<AcHost> acHostOption;
+        try {
+            acHostOption = acHostRepository.findByKey(clientKey);
+        } catch (Throwable throwable) {
+            return Promise.throwing(throwable);
+        }
+        final AcHost acHost = acHostOption.orElse(new Supplier<Option<AcHost>>() {
+            @Override
+            public Option<AcHost> get() {
+                try {
+                    return acHostRepository.findByUrl(baseUrl);
+                } catch (Throwable e) {
+                    return none(AcHost.class);
+                }
+            }
+        }).getOrElse(new AcHost());
+
+        acHost.setKey(clientKey);
+        acHost.setBaseUrl(baseUrl);
+        acHost.setPublicKey(publicKey);
+        acHost.setSharedSecret(sharedSecret);
+        acHost.setName(name);
+
         if (stripToNull(acHost.getPublicKey()) == null) {
             throw new InvalidAuthenticationRequestException("No public key provided in registration request");
         }
@@ -70,6 +104,16 @@ public class AcHostServiceImpl implements AcHostService {
         });
 
         return hostRegistered;
+    }
+
+    @Override
+    public Option<AcHost> findByKey(String consumerKey) throws Throwable {
+        return acHostRepository.findByKey(consumerKey);
+    }
+
+    @Override
+    public List<AcHost> all() throws Throwable {
+        return acHostRepository.all();
     }
 
 }