+++ project/Build.scala
@@ -5,7 +5,7 @@ import play.Project._
 object ApplicationBuild extends Build {
 
   val appName         = "ac-play-java"
-  val appVersion      = "0.9.0"
+  val appVersion      = "0.10.0-rc.1"
   val atlassianJwtVersion = "1.0.1"
 
   val appDependencies = Seq(