+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog - Play Java Module for Atlassian Connect
 
+## 0.10.0
+
+* Added support for pluggable persistence for AcHost data. Thanks to Alan Parkinson for the contribution
+* AcHost is now a class. AcHostModel has been removed.
+* Use core Ac logger for JWT. Thanks to Alan Parkinson for the contribution
+
 ## 0.9.0
 
 * Upgraded AUI to version 5.6.12